% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_uswds.R
\name{html_uswds}
\alias{html_uswds}
\title{HTML Output Based on the US Web Design Standards}
\usage{
html_uswds(highlight = NULL, css = NULL, keep_md = FALSE, ...)
}
\arguments{
\item{css}{Additional CSS file to be \strong{merged} in the document. It will
\strong{NOT} overwrite the CSS from the themes.}

\item{\dots}{Additional arguments passed to \code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This function is an alternative to the \code{\link[rmarkdown]{html_document}}
formats to create HTML document from R Markdown. It is intended to generate
single HTML pages that are based on the US Web Design Standards and primarily
scientific content. Attempts have been made to mirror a scientific journal
while maintaining the style design principles outlined in the web design
standards.
}
\examples{
\dontrun{
doc = system.file("rmarkdown", "templates", "usweb",
                  "skeleton", "skeleton.Rmd", package = "uswebr")
rmarkdown::render(doc, html_uswds(), output_dir = tempdir())
}
}
\author{
Josh M London
}
