% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_uswds_vignette.R
\name{html_uswds_vignette}
\alias{html_uswds_vignette}
\title{HTML Vignette Output Based on the US Web Design Standards}
\usage{
html_uswds_vignette(highlight = NULL, css = NULL, keep_md = FALSE,
  ...)
}
\arguments{
\item{css}{Additional CSS file to be \strong{merged} in the document. It will
\strong{NOT} overwrite the CSS from the themes.}

\item{\dots}{Additional arguments passed to \code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This function is a variation of \code{html_uswds}
formats to create HTML vignette from R Markdown. It is intended to generate
single HTML pages that are based on the US Web Design Standards and focussed
on package vignettes. Attempts have been made to reduce file size
while maintaining the style design principles outlined in the web design
standards.
}
\examples{
\dontrun{
doc = system.file("rmarkdown", "templates", "usweb",
                  "skeleton", "skeleton.Rmd", package = "uswebr")
rmarkdown::render(doc, html_uswds_vignette(), output_dir = tempdir())
}
}
\author{
Josh M London
}
