% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_report.R
\name{noaa_report}
\alias{noaa_report}
\title{Output a NOAA report.}
\usage{
noaa_report(toc = FALSE, css = NULL, fig_width = 8, ...)
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A NOAA report is a version of html_document with support for NOAA
colors, logos and based on the US Web Design Standards
}

