% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uswds_report.R
\name{uswds_report}
\alias{uswds_report}
\title{Output a USWDS scientific report (deprecated).}
\usage{
uswds_report(...)
}
\value{
a message about the function being deprecated
}
\description{
A USWDS report is a version of html_document with support for scientific
reports and based on the US Web Design Standards. This is now deprecated.
Use uswebr::html_uswds() instead
}
