% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uswds_vignette.R
\name{uswds_vignette}
\alias{uswds_vignette}
\title{Output a USWDS styled vignette}
\usage{
uswds_vignette(toc = FALSE, css = NULL, fig_width = 6, ...)
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A USWDS styled vignette is a lightweight version of html_document with based on
the US Web Design Standards
}

