% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usweb_color.R
\name{usweb_color}
\alias{usweb_color}
\title{Return color value (hex) for a color described in the US Web Design Standards}
\usage{
usweb_color(color_name)
}
\arguments{
\item{color_name}{Color name}
}
\description{
Return color value (hex) for a color described in the US Web Design Standards
}

