% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usweb_colors.R
\docType{data}
\name{usweb_main}
\alias{scale_color_usweb}
\alias{scale_colour_usweb}
\alias{scale_fill_usweb}
\alias{usweb_main}
\alias{usweb_pal}
\title{US Web Design Standards Color Palette (Discrete) and Scales}
\format{An object of class \code{character} of length 7.}
\usage{
usweb_main

usweb_pal()

scale_colour_usweb(...)

scale_color_usweb(...)

scale_fill_usweb(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
An 7-color qualitative discrete palette based on the UW Web Design Standards
and the colorblind-safe palette proposed by Okabe and Ito
}
\examples{
library("ggplot2")
library(scales)
show_col(usweb_pal()(7))
p <- ggplot(mpg, aes(class))
p + geom_bar(aes(fill=drv))
p + geom_bar(aes(fill=drv)) + uswebr_theme() + scale_fill_usweb()
}
\references{
\url{http://jfly.iam.u-tokyo.ac.jp/color}
\url{https://playbook.cio.gov/designstandards/visual-style/#palette}
}
\keyword{datasets}

