% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usweb_grad.R
\name{usweb_grad}
\alias{usweb_grad}
\title{Various color ramps based on the US Web Design Standards}
\usage{
usweb_grad(grad_name)
}
\arguments{
\item{pal_name}{Palette name}
}
\description{
Various color ramps based on the US Web Design Standards
}
\section{Palette names}{

\itemize{
  \item \code{primary}
  \item \code{primary_alt}
  \item \code{secondary}
  \item \code{gray}
  \item \code{gray_warm}
  \item \code{gold}
  \item \code{green}
  \item \code{cool_blue}
}
}

