% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usweb_palettes.R
\name{usweb_pal}
\alias{usweb_pal}
\alias{usweb_pal.cool_blue}
\alias{usweb_pal.gold}
\alias{usweb_pal.gray}
\alias{usweb_pal.gray_warm}
\alias{usweb_pal.green}
\alias{usweb_pal.main}
\alias{usweb_pal.primary}
\alias{usweb_pal.primary_alt}
\alias{usweb_pal.secondary}
\title{Various color palettes based on the US Web Design Standards}
\usage{
usweb_pal(n)

\method{usweb_pal}{main}(n = 5)

\method{usweb_pal}{primary}(n = 4)

\method{usweb_pal}{primary_alt}(n = 4)

\method{usweb_pal}{secondary}(n = 4)

\method{usweb_pal}{gray}(n = 5)

\method{usweb_pal}{gray_warm}(n = 2)

\method{usweb_pal}{gold}(n = 4)

\method{usweb_pal}{green}(n = 4)

\method{usweb_pal}{cool_blue}(n = 4)
}
\arguments{
\item{n}{number of colors}
}
\description{
Various color palettes based on the US Web Design Standards
}
\section{Methods (by class)}{
\itemize{
\item \code{main}: Main color palette (categorical, 5 colors)

\item \code{primary}: Primary color palette (gradient, 4 colors)

\item \code{primary_alt}: Alternate primary color palette (gradient, 4 colors)

\item \code{secondary}: Secondary color palette (gradient, 4 colors)

\item \code{gray}: Gray color palette (gradient, 5 colors)

\item \code{gray_warm}: Warm gray color palette (gradient, 2 colors)

\item \code{gold}: Gold color palette (gradient, 4 colors)

\item \code{green}: Green color palette (gradient, 4 colors)

\item \code{cool_blue}: Cool blue color palette (gradient, 4 colors)
}}

