% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uswds_report.R
\name{uswds_report}
\alias{uswds_report}
\title{Output a USWDS scientific report.}
\usage{
uswds_report(toc = FALSE, css = NULL, fig_width = 8, ...)
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
A USWDS report is a version of html_document with support for scientific
reports and based on the US Web Design Standards
}
