% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usweb_colors.R
\docType{data}
\name{usweb_colors}
\alias{usweb_colors}
\title{Colors described within the U.S. Web Design Standards}
\format{A named vector with colors and hex values}
\source{
\url{https://playbook.cio.gov/designstandards/visual-style/#colors}
}
\usage{
usweb_colors
}
\description{
Colors described within the U.S. Web Design Standards
}
\keyword{datasets}

